+++ src/Sentry/Internal/IStringOrRegexMatcher.cs
@@ -1,0 +1,6 @@
+namespace Sentry.Internal;
+
+internal interface IStringOrRegexMatcher
+{
+    bool IsMatch(StringOrRegex stringOrRegex, string value);
+}